✨ Add database in io module
Made so database configuration can be streamlined & handled the same way in all projects. Important properties are added there. It's highly configurable so you don't have to use the same structure in the config. The ?allowMultiQueries=true is needed so you can pipe sql files with multiple statements in the driver.
Also our license got added, so it's auto update.